% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cigarette_sales-data.R
\docType{data}
\name{cigarette_sales}
\alias{cigarette_sales}
\title{Cigarette sales and retail prices by state}
\format{
Each row is a year, each column is either cigarette sales or retail cigarette price by state. State is indicated by suffix with FIPS code.
}
\usage{
data(cigarette_sales)
}
\description{
Data from Abadie, Diamond, and Hainmuelle (2010)
}
\examples{
data(cigarette_sales)
mean(cigarette_sales$year)
mean(cigarette_sales$cigsale_6)
}
\references{
Abadie, Alberto, Alexis Diamond, and Jens Hainmueller. 2010. “Synthetic Control Methods for Comparative Case Studies: Estimating the Effect of California’s Tobacco Control Program.” Journal of the American Statistical Association 105 (490): 493–505. https://doi.org/10.1198/jasa.2009.ap08746.
}
\keyword{datasets}
