% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvexHullCheck.R
\name{ConvexHullCheck}
\alias{ConvexHullCheck}
\title{Check if synthetic variable is in convex hull}
\usage{
ConvexHullCheck(
  coef.exact = SCUL.output$coef.exact,
  x.DonorPool = SCUL.input$x.DonorPool,
  y.scul = SCUL.output$y.scul,
  y.actual = SCUL.output$y.actual,
  TreatmentBeginsAt = SCUL.output$TreatmentBeginsAt
)
}
\arguments{
\item{coef.exact}{The sparse matrix of coefficients used to construc the synthetic estimate. 
Default is SCUL.output$coef.exact.}

\item{x.DonorPool}{The (T by K) donor pool matrix. Default is SCUL.input$x.DonorPool.}

\item{y.scul}{Synthetic data created by SCUL procedure. Default is SCUL.output$y.scul.}

\item{y.actual}{The actual (target) data. Default is SCUL.output$y.actual.}

\item{TreatmentBeginsAt}{Default is  SCUL.output$TreatmentBeginsAt.}
}
\value{
indicator A binary indicator if scul prediction is in convex hull
}
\description{
Check if synthetic variable is in convex hull.
}
