% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePlaceboDistribution.R
\name{CreatePlaceboDistribution}
\alias{CreatePlaceboDistribution}
\title{Run scul procedure on placebo pool to create null distribution}
\usage{
CreatePlaceboDistribution(
  x.PlaceboPool = SCUL.input$x.PlaceboPool,
  TreatmentBeginsAt = SCUL.input$TreatmentBeginsAt,
  PostPeriodLength = nrow(SCUL.input$y) - SCUL.input$TreatmentBeginsAt + 1,
  PrePeriodLength = SCUL.input$TreatmentBeginsAt - 1,
  NumberInitialTimePeriods = SCUL.input$NumberInitialTimePeriods,
  OutputFilePath = SCUL.input$OutputFilePath,
  CohensDThreshold = SCUL.input$CohensDThreshold,
  TrainingPostPeriodLength = SCUL.input$TrainingPostPeriodLength
)
}
\arguments{
\item{x.PlaceboPool}{A (T by J) data frame containing all products that you wish to include in the placebo distribution
Must be sorted by time. Default is SCUL.input$x.PlaceboPool.}

\item{TreatmentBeginsAt}{An integer indicating which row begins treatment. Default is  SCUL.output$TreatmentBeginsAt.}

\item{PostPeriodLength}{An integer that indicates the length of the post-treatment period.
Defailt is calculated using SCUL.input data.}

\item{PrePeriodLength}{An integer that indicates the length of the pre-treatment period.
Defailt is calculated using SCUL.input data.}

\item{NumberInitialTimePeriods}{An integer that indicates the number of time periods desired in the training data for the first cross-validation run.
Default is the stated amount in SCUL.input data.}

\item{OutputFilePath}{Output file path. Default is  SCUL.input$OutputFilePath.}

\item{CohensDThreshold}{A real number greater than 0, indicating the Cohen's D threshold at which
fit is determined to be "poor". The difference is in standard deviation units. Default is SCUL.input$CohensDThreshold.}

\item{TrainingPostPeriodLength}{The number of timer periods post-treatment for training data. Defaults to all time since treatment begins.}
}
\value{
list  A list of standardized placbo effect sizes
}
\description{
Runs the scul procedure on the whole placebo pool
(using itself as the donor pool; this could be altered to have a custom donor pool)
to create null distribution that can be used for inference in later steps.
}
