% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistogramOfDiff.R
\name{HistogramOfDiff}
\alias{HistogramOfDiff}
\title{Make histogram of null with effect size}
\usage{

  HistogramOfDiff(y.StandardizedDifference = SCUL.pvalue$y.StandardizedDifference,
  y.placebo.StandardizedDifference = SCUL.inference$y.placebo.StandardizedDifference,
  time = SCUL.input$time,
  TreatmentBeginsAt = SCUL.input$TreatmentBeginsAt,
  OutputFilePath = SCUL.input$OutputFilePath, width = 40,
  widthBreaks = c(-width, -width/2, 0, width/2, width), height = 0.05,
  AdjustmentParm = 3, BandwidthParm = 3)
}
\arguments{
\item{y.StandardizedDifference}{The standardized difference between target data and synthetic data series across time. Default is SCUL.inference$y.StandardizedDifference.}

\item{y.placebo.StandardizedDifference}{Standardized difference between each placebo data series and its synthetic prediction across time. Default is SCUL.inference$y.placebo.StandardizedDifference.}

\item{time}{The time variable across. Default is SCUL.input$time.}

\item{TreatmentBeginsAt}{An integer indicating which row begins treatment. Default is  SCUL.output$TreatmentBeginsAt.}

\item{OutputFilePath}{Output file path. Default is SCUL.input$OutputFilePath.}

\item{width}{width of histogram. Default is 20}

\item{widthBreaks}{width breaks histogram. Default is 5, evenly spaced.}

\item{height}{height of histogram. Default is 0.05}

\item{AdjustmentParm}{Adjustment parameter for guassian histogram. Default is 3.}

\item{BandwidthParm}{Bandwidth parameter for guassian histogram. Default is 3.}
}
\value{
graphs  Two histograms with effect size
}
\description{
Plot a histogram of the null two different ways with estimated effect size plotted as well.
}
