% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysd.R
\name{mysd}
\alias{mysd}
\title{Standardize column vector}
\usage{
mysd(z)
}
\arguments{
\item{z}{Input}
}
\description{
Function to standardize a column vector in a similar way that glmnet does. Used to calculate max penalty in \code{lambda} grid.
Taken from: https://stackoverflow.com/questions/23686067/default-lambda-sequence-in-glmnet-for-cross-validation
}
