% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotShareForFirst.R
\name{PlotShareForFirst}
\alias{PlotShareForFirst}
\title{Graph output from SCUL procedure}
\usage{
PlotShareForFirst(
  coef.exact = SCUL.output$coef.exact,
  x.DonorPool = SCUL.input$x.DonorPool,
  y.scul = SCUL.output$y.scul,
  y.actual = SCUL.output$y.actual,
  time = SCUL.output$time,
  OutputFilePath = SCUL.input$OutputFilePath,
  TreatmentBeginsAt = SCUL.output$TreatmentBeginsAt
)
}
\arguments{
\item{coef.exact}{The sparse matrix of coefficients used to construc the synthetic estimate. 
Default is SCUL.output$coef.exact.}

\item{x.DonorPool}{The (T by K) donor pool matrix. Default is SCUL.input$x.DonorPool.}

\item{y.scul}{Synthetic data created by SCUL procedure. Default is SCUL.output$y.scul.}

\item{y.actual}{The actual (target) data. Default is SCUL.output$y.actual.}

\item{time}{Default is  SCUL.output$time.}

\item{OutputFilePath}{Output file path. Default is  SCUL.input$OutputFilePath.}

\item{TreatmentBeginsAt}{An integer indicating which row begins treatment. Default is  SCUL.output$TreatmentBeginsAt.}
}
\value{
graphs  Four plots that show contribution of each donor to synthetic estimate and a similar table.
}
\description{
Plot output from the SCUL procedure in levels, differences, and standardized difference
Must also have run this command. webshot::install_phantomjs()
}
