% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{Preprocess}
\alias{Preprocess}
\title{Process and clean input data for SCUL procedure}
\usage{
Preprocess(z)
}
\arguments{
\item{z}{A matrix or a data frame.}
}
\value{
A cleaned matrix or data frame with violating columns dropped.
}
\description{
This function loads a matrix or a data-fram a file as a matrix.
It ensures that each column of the matrix is numeric, has all non-missing values,
and is not constant. It drops any columns that violate these conditions.
}
