% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessSubset.R
\name{PreprocessSubset}
\alias{PreprocessSubset}
\title{Drops columns that have no variation in training subsets}
\usage{
PreprocessSubset(
  z,
  TreatmentBeginsAt,
  NumberInitialTimePeriods = nrow(z) - TreatmentBeginsAt + 1,
  PostPeriodLength = nrow(z) - TreatmentBeginsAt + 1,
  PrePeriodLength = TreatmentBeginsAt - 1,
  TrainingPostPeriodLength = nrow(z) - TreatmentBeginsAt + 1
)
}
\arguments{
\item{z}{A matrix or a data frame.}

\item{TreatmentBeginsAt}{Indicator of when (which row) treatment begins.}

\item{NumberInitialTimePeriods}{The number of initial time periods in analysis. Defaults to length of post-treatment period.}

\item{PostPeriodLength}{The number of timer periods post-treatment. Defaults to all time since treatment begins.}

\item{PrePeriodLength}{The number of time periods before treatment (i.e. the length of the entire training dataset). Defaults to time before treatment begins.}

\item{TrainingPostPeriodLength}{The number of timer periods post-treatment for training data. Defaults to all time since treatment begins.}
}
\value{
A cleaned matrix or data frame with violating columns dropped.
}
\description{
This function loads a matrix or a data-fram a file as a matrix.
It ensures that for all relevant subsets that could be used as a training 
dataset for the later SCUL procedure have variation. 
It drops any columns that violate these conditions.
}
