% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PValue.R
\name{PValue}
\alias{PValue}
\title{Calculate a simple bounded p-value for the treatment effect}
\usage{
PValue(
 
    x.PlaceboPool.StandardizedDiff.trimmed = SCUL.inference$y.placebo.StandardizedDifference,
  TreatmentBeginsAt = SCUL.input$TreatmentBeginsAt,
  OutputFilePath = SCUL.input$OutputFilePath,
  CohensDThreshold = SCUL.input$CohensDThreshold,
  y.actual = SCUL.output$y.actual,
  y.scul = SCUL.output$y.scul
)
}
\arguments{
\item{x.PlaceboPool.StandardizedDiff.trimmed}{A (T by L), where L<=J)  data frame containing all products that are included in the placebo distribution
Trimmed by Cohen's D using CreatePlaceboDistribution function. Default is SCUL.inference$x.PlaceboPool.StandardizedDiff.trimmed}

\item{TreatmentBeginsAt}{An integer indicating which row begins treatment. Default is  SCUL.output$TreatmentBeginsAt.}

\item{OutputFilePath}{Output file path. Default is  SCUL.input$OutputFilePath.}

\item{CohensDThreshold}{A real number greater than 0, indicating the Cohen's D threshold at which
fit is determined to be "poor". The difference is in standard deviation units. Default is SCUL.input$CohensDThreshold.}

\item{y.actual}{The actual (target) data. Default is SCUL.output$y.actual.}

\item{y.scul}{Synthetic data created by SCUL procedure. Default is SCUL.output$y.scul.}

\item{PostPeriodLength}{An integer that indicates the length of the post-treatment period.
Defailt is calculated using SCUL.input data.}

\item{PrePeriodLength}{An integer that indicates the length of the pre-treatment period.
Defailt is calculated using SCUL.input data.}

\item{NumberInitialTimePeriods}{An integer that indicates the number of time periods desired in the training data for the first cross-validation run. 
Default is the stated amount in SCUL.input data.}
}
\value{
list  The bounds for the rank based p-value based upon rank of mean absolute value of post-treatment standardized effect against null distribution.
}
\description{
Determines the rank of the mean absolute value of the standardized treatment effect 
of the target product relative to the placebo distribution. There are many alternative 
ways to use the placebo distribution to calculate a mean value (e.g., mean p-value or another test statistic). 
This is a simple example of one such way.
}
