% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PValue.R
\name{PVaule}
\alias{PVaule}
\title{Run placebo scul procedure on placebo pool and create p-values}
\usage{
PVaule(x.PlaceboPool = SCUL.input$x.PlaceboPool,
  TreatmentBeginsAt = SCUL.input$TreatmentBeginsAt,
  PostPeriodLength = nrow(SCUL.input$y) - SCUL.input$TreatmentBeginsAt +
  1, PrePeriodLength = SCUL.input$TreatmentBeginsAt - 1,
  NumberInitialTimePeriods = SCUL.input$NumberInitialTimePeriods,
  OutputFilePath = SCUL.input$OutputFilePath,
  CohensDThreshold = SCUL.input$CohensDThreshold,
  y.actual = SCUL.output$y.actual, y.scul = SCUL.output$y.scul)
}
\arguments{
\item{x.PlaceboPool}{A (T by J) data frame containing all products that you wish to include in the placebo distribution
Must be sorted by time. Default is SCUL.input$x.PlaceboPool.}

\item{TreatmentBeginsAt}{An integer indicating which row begins treatment. Default is  SCUL.output$TreatmentBeginsAt.}

\item{PostPeriodLength}{An integer that indicates the length of the post-treatment period.
Defailt is calculated using SCUL.input data.}

\item{PrePeriodLength}{An integer that indicates the length of the pre-treatment period.
Defailt is calculated using SCUL.input data.}

\item{NumberInitialTimePeriods}{An integer that indicates the number of time periods desired in the training data for the first cross-validation run. 
Default is the stated amount in SCUL.input data.}

\item{OutputFilePath}{Output file path. Default is  SCUL.input$OutputFilePath.}

\item{CohensDThreshold}{A real number greater than 0, indicating the Cohen's D threshold at which
fit is determined to be "poor". The difference is in standard deviation units. Default is SCUL.input$CohensDThreshold.}

\item{y.actual}{The actual (target) data. Default is SCUL.output$y.actual.}

\item{y.scul}{Synthetic data created by SCUL procedure. Default is SCUL.output$y.scul.}
}
\value{
list  A list of standardized placbo effect sizes and bounds for the p-value.
}
\description{
Runs the same scul procedure on every column in the donor pool to create
a rank based p-value.
}
