% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmokePlot.R
\name{SmokePlot}
\alias{SmokePlot}
\title{Make smoke plot}
\usage{

  SmokePlot(y.StandardizedDifference = SCUL.inference$y.StandardizedDifference,
  y.placebo.StandardizedDifference = SCUL.inference$y.placebo.StandardizedDifference,
  time = SCUL.input$time, OutputFilePath = SCUL.input$OutputFilePath)
}
\arguments{
\item{y.StandardizedDifference}{The standardized difference between target data and synthetic data series across time. Default is SCUL.inference$y.StandardizedDifference.}

\item{y.placebo.StandardizedDifference}{Standardized difference between each placebo data series and its synthetic prediction across time. Default is SCUL.inference$y.placebo.StandardizedDifference.}

\item{time}{The time variable across. Default is SCUL.input$time.}

\item{OutputFilePath}{Output file path. Default is SCUL.input$OutputFilePath.}
}
\value{
graph  A smoke plot of the standardized effect size compared to placbos.
}
\description{
Plot standardized differences of all placebo goods and target good.
}
