% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PValue.R
\name{PValue}
\alias{PValue}
\title{Calculate a simple bounded p-value for the treatment effect reporting the larger end of the bound}
\usage{
PValue(
  x.PlaceboPool.full = SCUL.inference$y.placebo.StandardizedDifference.Full,
  x.PlaceboPool.CohensD = SCUL.inference$y.placebo.CohensD,
  TreatmentBeginsAt = SCUL.input$TreatmentBeginsAt,
  OutputFilePath = SCUL.input$OutputFilePath,
  CohensD = SCUL.input$CohensDThreshold,
  y.actual = SCUL.output$y.actual,
  y.scul = SCUL.output$y.scul,
  StartTime = SCUL.input$TreatmentBeginsAt,
  EndTime = nrow(SCUL.output$y.scul)
)
}
\arguments{
\item{x.PlaceboPool.full}{A (T by L), where L<=J)  data frame containing all products that are included in the placebo distribution
Default is SCUL.inference$y.placebo.StandardizedDifference.Full}

\item{x.PlaceboPool.CohensD}{A (1 by L)  data frame containing all pre-period Cohen's D fit statistic for each placebo unit.
Default is SCUL.inference$y.placebo.CohensD,}

\item{TreatmentBeginsAt}{An integer indicating which row begins treatment. Default is  SCUL.output$TreatmentBeginsAt.}

\item{OutputFilePath}{Output file path. Default is  SCUL.input$OutputFilePath.}

\item{CohensD}{A real number greater than 0, indicating the Cohen's D threshold at which
fit is determined to be "poor". The difference is in standard deviation units. Default is SCUL.input$CohensDThreshold.}

\item{y.actual}{The actual (target) data. Default is SCUL.output$y.actual.}

\item{y.scul}{Synthetic data created by SCUL procedure. Default is SCUL.output$y.scul.}

\item{StartTime}{The begining time period for which the average pseduo treatment effect is calculated. T}

\item{EndTime}{The end time period for which the average pseduo treatment effect is calculated. T}
}
\value{
list  The bounds for the rank based p-value based upon rank of mean absolute value of post-treatment standardized effect against null distribution.
}
\description{
Determines the rank of the mean absolute value of the standardized treatment effect
of the target product relative to the placebo distribution. There are many alternative
ways to use the placebo distribution to calculate a mean value (e.g., mean p-value or another test statistic).
This is a simple example of one such way.
}
